// src/api/auth/auth.errors.enum.ts
export enum AuthErrors {
  LOGIN_FAILED = 'AUTH.LOGIN_FAILED',
  SIGNUP_FAILED = 'AUTH.SIGNUP_FAILED',
  LOGOUT_FAILED = 'AUTH.LOGOUT_FAILED',
  EMAIL_WAS_NOT_SENT = 'AUTH.EMAIL_WAS_NOT_SENT',
  ALREADY_EXISTS = 'AUTH.ALREADY_EXISTS',
  NOT_EXISTS = 'AUTH.NOT_EXISTS',
  UNAUTHORIZED = 'AUTH.UNAUTHORIZED',
  CHANGE_PASSWORD_FAILED = 'AUTH.CHANGE_PASSWORD_FAILED',
  EXPIRED = 'AUTH.EXPIRED',
}